requirements.txt
streamlit==1.45.1
sentence-transformers==4.1.0
faiss-cpu==1.11.0
numpy==1.26.4
Pillow==10.2.0